Move validation logic to separate file
